Project setup with deployment target 11.0
