+++ CodingAssignment/CodingAssignment/AppDelegates/AppDelegate.swift
@@ -1,0 +1,41 @@
+//
+//  AppDelegate.swift
+//  CodingAssignment
+//
+//  Created by Lokesh Jain on 21/07/20.
+//  Copyright © 2020 Lokesh. All rights reserved.
+//
+
+import UIKit
+
+@UIApplicationMain
+class AppDelegate: UIResponder, UIApplicationDelegate {
+  
+  var window: UIWindow?
+  
+  func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+    
+    // Check iOS before configuring UIWindow
+    if #available(iOS 13.0, *) {
+      // Set rootViewController in SceneDelegate
+    } else {
+      // Configuring UIWindow and setting its rootViewController
+      let window = UIWindow(frame: UIScreen.main.bounds)
+      window.rootViewController = UINavigationController(rootViewController: ListController()) // Our initial view controller.
+      window.makeKeyAndVisible()
+      self.window = window
+    }
+    return true
+  }
+  
+  // MARK: UISceneSession Lifecycle
+  @available(iOS 13.0, *)
+  func application(_ application: UIApplication, configurationForConnecting connectingSceneSession: UISceneSession, options: UIScene.ConnectionOptions) -> UISceneConfiguration {
+    return UISceneConfiguration(name: "Default Configuration", sessionRole: connectingSceneSession.role)
+  }
+  
+  @available(iOS 13.0, *)
+  func application(_ application: UIApplication, didDiscardSceneSessions sceneSessions: Set<UISceneSession>) {
+  }
+}
+
