+++ CodingAssignment/CodingAssignment/AppDelegates/SceneDelegate.swift
@@ -1,0 +1,42 @@
+//
+//  SceneDelegate.swift
+//  CodingAssignment
+//
+//  Created by Lokesh Jain on 21/07/20.
+//  Copyright © 2020 Lokesh. All rights reserved.
+//
+
+import UIKit
+
+@available(iOS 13.0, *)
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+  
+  var window: UIWindow?
+  
+  func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+    guard let windowScene = (scene as? UIWindowScene) else { return }
+    
+    // Setting rootViewController on UIWindow for 13.0 devices
+    let window = UIWindow(windowScene: windowScene)
+    window.rootViewController = UINavigationController(rootViewController: ListController()) // Our initial view controller.
+    window.makeKeyAndVisible()
+    self.window = window
+  }
+  
+  func sceneDidDisconnect(_ scene: UIScene) {
+  }
+  
+  func sceneDidBecomeActive(_ scene: UIScene) {
+  }
+  
+  func sceneWillResignActive(_ scene: UIScene) {
+  }
+  
+  func sceneWillEnterForeground(_ scene: UIScene) {
+  }
+  
+  func sceneDidEnterBackground(_ scene: UIScene) {
+    
+  }
+}
+
