+++ CodingAssignment/CodingAssignment/CustomViews/ListCell.swift
@@ -1,0 +1,21 @@
+//
+//  ListCell.swift
+//  CodingAssignment
+//
+//  Created by Lokesh Jain on 21/07/20.
+//  Copyright © 2020 Lokesh. All rights reserved.
+//
+
+import UIKit
+
+class ListCell: UITableViewCell {
+  
+  override func awakeFromNib() {
+    super.awakeFromNib()
+  }
+  
+  override func setSelected(_ selected: Bool, animated: Bool) {
+    super.setSelected(selected, animated: animated)
+    
+  }
+}
