+++ CodingAssignment/CodingAssignment/Reachability/ReachabilityManager.swift
@@ -1,0 +1,9 @@
+//
+//  ReachabilityManager.swift
+//  CodingAssignment
+//
+//  Created by Lokesh Jain on 21/07/20.
+//  Copyright © 2020 Lokesh. All rights reserved.
+//
+
+import Foundation
