--- CodingAssignment/CodingAssignment/ViewController.swift
@@ -1,20 +1,0 @@
-//
-//  ViewController.swift
-//  CodingAssignment
-//
-//  Created by Lokesh Jain on 21/07/20.
-//  Copyright © 2020 Lokesh. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
