+++ CodingAssignment/CodingAssignment/ViewControllers/ListController.swift
@@ -1,0 +1,18 @@
+//
+//  ListController.swift
+//  CodingAssignment
+//
+//  Created by Lokesh Jain on 21/07/20.
+//  Copyright © 2020 Lokesh. All rights reserved.
+//
+
+import UIKit
+
+class ListController: UIViewController {
+  
+  override func viewDidLoad() {
+    super.viewDidLoad()
+    view.backgroundColor = .white
+  }
+}
+
