+++ CodingAssignment/CodingAssignment/ViewModels/ListViewModel.swift
@@ -1,0 +1,9 @@
+//
+//  ListViewModel.swift
+//  CodingAssignment
+//
+//  Created by Lokesh Jain on 21/07/20.
+//  Copyright © 2020 Lokesh. All rights reserved.
+//
+
+import Foundation
